refactor(api): extract shared OpenAPI tag/schema filter (DRY)
collectRefs, pruneSchemas, filterByTags, and HTTP_METHODS were duplicated
between main.ts and scripts/generate-openapi.ts. Canonical implementations
now live in api/src/common/openapi-filter.ts and are imported by both files.
The generate script's filterByTags now takes an explicit Set<string> tags arg
(matching the main.ts signature) instead of closing over a module-level const.
No behavior change: openapi.json no-drift confirmed, docs split unchanged,
73/73 tests green.
